backend/requirements.txt
ultralytics==8.4.33
opencv-python==4.13.0.92
numpy==2.2.6
fastapi==0.127.0
uvicorn[standard]==0.40.0
confluent-kafka==2.14.0
psycopg2-binary==2.9.9
